Fix handling of modelObject in log messages and updae location of ruleset map.
When running arelle in certain web server modes, the logging cannot
handle modelObjects.  Updated to detect this situation and use the
sourceFileLine on the log instead of modelObject.

Updated the ruleset map to point to the version of the DQC rules in the
release branch. Updated the readme.md file for typs and to include
further documentation about updating the ruleset map for future
releases.