v6 - Update for duplicate non neg items and SHE Transition items (#350)
* Updated 2018 taxonomy rules to incorporate transition elements

There are some companies that haven't adopted ASU 2017-12 as of yet. This guidance is effective for annual periods beginning after December 15, 2018. It appears that perhaps those companies are supposed to use "Transition" concepts under us-gaap:Asu201712TransitionAbstract, which includes members that are not structured in a hypercube in the 2018 taxonomy, but were in the previous taxonomies.

* Updated compiled code under the same version number

* Update to remove duplicate DQC_0015 non-neg elements (documentation)