Updated Rules for Identified Changes
Update rule DQC.US.0067 to allow for more 440 exception items specifically the following:
IncreaseDecreaseInDeferredRevenue
DeferredRevenueAdditions
DeferredRevenueRevenueRecognized1
DeferredRevenueRefundPayments1
DeferredRevenuePeriodIncreaseDecrease
DeferredRevenueByArrangementDisclosureTextBlock

Updated rule DQC.US.0076 to only require alignment of the satisfaction Date axis with the obligation value.  This value does not have to have matching dimensions between the values for RevenueRemainingPerformanceObligationExpectedTimingOfSatisfactionPeriod1 and RevenueRemainingPerformanceObligation or RevenueRemainingPerformanceObligationPercentage
For example: In some cases a filer may report the  obligation or percentage value with a maximum member or a minimum member using the range axis. The way the rule was previously written  required that the satisfaction period element would also need to be tagged twice using the minimum member and maximum member. The rule now allows these the obligation and the satisfaction period of that obligation to have different dimensional qualifications.

The rule was also updated for 2018 to not require period alignment. The satisfaction period context date no longer has tp match the context date of the reported data.

Updated rule DQC.US.0071 to ignore facts that use the InitialApplicationPeriodCumulativeEffectTransitionAxis.  This axis can break up changes by product and will not appear on the income statement. If revenue is broken down by product and revenue recognition adjustments by product, then the table may only show a single product line as that was the only product that an adjustment applied to. In this case the rule would produce an inaccurate error. The rule was updated to ignore any tables that include the axis InitialApplicationPeriodCumulativeEffectTransitionAxis.