Updated to handle inconsistent decimals
When getting values for the FS Arell will eliminate duplicate values, however in some cases duplicates are allowed where decimals are different. This is an issue as both values get included in the total. The value with the inappropriate decimals must be removed. A function was added to identify the decimals of the financial statements and then only values with these decimals are included in the calculation check.