Updated rule 118 to handle the consolidated items axis correctly
Updated:
If the consolidated axis is not in the notes then we can consider using. If in notes and on the table hybercube then the rule stops running, unless processing the current context which is the consolidated items axis. If not in notes then check the addend does not include the dimension. If it does then we stop processing the rule.