Update to rule 118 to fix performance
This is a major overhaul of 118 to fix issues when have lots of dimensions and facts.  The rule is changed to do dependent processing of facts to avoid calculating facts in bulk and then matching them. This means pushing the context of the sum into the component parts fact selection.