Updated rule 195 to add Materiality
After a number of complaints we updated rule DQC.US.0195.10627 to have a materiality check.  This rule looks at line items in SHE that should not impact retained earnings. Because multiple transactions are collapsed into a line  for grants and weird plans that pay dividends and when restricted they are reversed there can be a residual impact on retained earnings.  This generally should be a different line item, but companies complain that this is immaterial and don't want to change their reports. To facilitate this we check the total impact of the line on shareholders equity and only flag an error if the retained earnings impact is more than 10% of the total impact on SHE. If they get an error they should look to tag the retained earnings portion with a different value.